refactor(range): use LSP definition of a range (almost)
I created the Range object to get rid of the juggling between LSP and Neovim ranges.
I was only partially satisfied with the result, the complexity was contained but I introduced yet another definition of
range which kind of defies the purpose.

So I decided to stick to the LSP definition which is the only constant since on Neovim sides different APIs have
different definitions.
The only concession I made was to use byte position instead of character position. Mainly because the ranges will be
used to select text in a buffer which uses bytes.

But I'm still considering the option to use character position instead since it's will be represented the same way
without having to know about the encoding.
This way it's up to the code using the range to know about it and since it will have to select or retrieve matching text
or things like that it's most likely that it will have the necessary information to find out this encoding.