refactor(tests): improve init_spec.lua to be more resilient
I don't mock the `get_client()` function anymore, instead I use the real implementation of `client.select()` and
configure my fake client to provide the necessary information for the selection to work.
This way the test does not depends on a fake implementation and allow me to have a more realistic feedback when I run
it.